improve event details sections
